// src/lib2/models/PeriodeComptableDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PeriodeComptableDto = {
    id?: string;
    code: string;
    cloturee: boolean;
    notes?: string;
    exercice_id?: string;
    dateDebut: string;
    dateFin: string;
    dateCloture?: string;
    createdAt?: string;
    updatedAt?: string;
    createdBy?: string;
    updatedBy?: string;
};